Document part order field and correct behavior on create (not delete yet).
