+++ src/databaseHelpers/documents.js
@@ -3,7 +3,7 @@ const createStandardFunctions = require('./createStandardFunctions');
 
 const standardFunctions = createStandardFunctions(
   'documents',
-  ['id', 'name', 'visibility']
+  ['id', 'name', 'visibility', 'partOrder']
 );
 
 module.exports = standardFunctions;
